# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Todo:
    properties:
      authorId:
        type: integer
      category:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo Management API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login an user with the input paylod
      parameters:
      - description: User successfully login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/User'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Login an user
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Register an user with the input paylod
      parameters:
      - description: User successfully registered
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/User'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Register an user
      tags:
      - users
  /todos:
    get:
      consumes:
      - application/json
      description: Get details of all todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Get details of all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo with the input paylod
      parameters:
      - description: Todo has been created
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/Todo'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Create a new todo
      tags:
      - todos
  /todos/{id:
    delete:
      consumes:
      - application/json
      description: Delete a todo with the input paylod
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete a todo
      tags:
      - todos
  /todos/{id}:
    get:
      consumes:
      - application/json
      description: Get details of todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Get details of todo
      tags:
      - todo
    put:
      consumes:
      - application/json
      description: Update a todo with the input paylod
      parameters:
      - description: Todo has been created
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/Todo'
          type: object
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Update a todo
      tags:
      - todos
swagger: "2.0"